AORTA-17 CLI command for report generation : refactored cli (#71)
